+++ src/render.c
@@ -55,7 +55,7 @@ void flush_window(f_mtdt* Buff)
 	fflush(stdout);
 }
 
-void upper_bar(f_mtdt* Buff, ui_mtdt Ui)
+void upper_bar(f_mtdt* Buff, win_mtdt Ui)
 {
 	const char* logo_half = "|`` \0"; // Lower and upper parts are the same.
 	const char* dots      = "[...]\0";
@@ -67,10 +67,10 @@ void upper_bar(f_mtdt* Buff, ui_mtdt Ui)
 	term_t fname_max =
 	get_term_sz(Buff, 'X') - (term_t) (strlen(logo_half) + strlen(dots));
 
-	ANSI_BOLD();
+	ANSI_INVERT();
 
-	/* Sometimes the empty space of width Ui.line_num_len will rendered before the
-	upper bar. Adding the carriage return before it fixes the problems. Just
+	/* Sometimes the empty space of width Ui.line_num_len will rendered before
+	the upper bar. Adding the carriage return before it fixes the problems. Just
 	handling with terminals' quirk modes. For any other output of the program CR
 	is not necessary, eg. for errors messages. They can be shifted. */
 	printf("\r%s", logo_half);
@@ -78,7 +78,8 @@ void upper_bar(f_mtdt* Buff, ui_mtdt Ui)
 	if(strlen(Buff->fname) <= fname_max)
 	{
 		// Whole filename will be displayed.
-		printf("%s\n", Buff->fname);
+		printf("%s%*s\n", Buff->fname, get_term_sz(Buff, 'X')
+		- (term_t) (strlen(logo_half) + strlen(Buff->fname)), " ");
 	}
 	else
 	{
@@ -91,26 +92,28 @@ void upper_bar(f_mtdt* Buff, ui_mtdt Ui)
 
 	if((ACT_LN_LEN < Ui.text_x) || ((ACT_LN_LEN - Buff->cusr_x) < Ui.text_x))
 	{
-		printf("%*d^\n", indicator_width,
+		printf("%*d^ \n", indicator_width,
 		get_term_sz(Buff, 'X') - Ui.line_num_len - SPACE_SZ);
 	}
 	else if((ACT_LN_LEN - Buff->cusr_x) >= Ui.text_x)
 	{
-		printf("%*d^\n", indicator_width, ACT_LN_LEN - Buff->cusr_x);
+		printf("%*d^ \n", indicator_width, ACT_LN_LEN - Buff->cusr_x);
 	}
 	ANSI_RESET();
 }
 
-void lower_bar(void)
+void lower_bar(f_mtdt* Buff)
 {
-	ANSI_BOLD();
+ANSI_INVERT();
 
-	printf("\n%s", LBAR_STR);
+	printf("\n%s%*s",
+	LBAR_STR, get_term_sz(Buff, 'X') - TERM_X_MIN + SPACE_SZ, " ");
 
 	ANSI_RESET();
+
 }
 
-void scroll_line_x(f_mtdt* Buff, ui_mtdt Ui)
+void scroll_line_x(f_mtdt* Buff, win_mtdt Ui)
 {
 	_Bool mv_right = 0;
 
@@ -147,14 +150,14 @@ buff_t scroll_lines(f_mtdt* Buff)
 
 void print_line_num(buff_t line, uint8_t line_num_len)
 {
-	ANSI_BOLD();
+	ANSI_INVERT();
 	printf("%*d", line_num_len - SPACE_SZ, line + INDEX);
 
 	ANSI_RESET();
 	putchar(' ');
 }
 
-void display_text(f_mtdt* Buff, ui_mtdt Ui)
+void display_text(f_mtdt* Buff, win_mtdt Ui) // TODO: CUSR_Y PROPER RENDERING.
 {
 	// Previous lines. If scrolled. Only beginning is shown.
 	for(buff_t line = scroll_lines(Buff); line < Buff->lines; line++)
@@ -205,7 +208,7 @@ void fill(f_mtdt* Buff)
 
 void window(f_mtdt* Buff)
 {
-	ui_mtdt Ui;
+	win_mtdt Ui;
 
 	// Snprinf isn't needed because the format specifier gives a warning.
 	sprintf(Ui.line_num_str, "%u", Buff->lines + INDEX);
@@ -220,11 +223,11 @@ void window(f_mtdt* Buff)
 	display_text(Buff, Ui);
 	fill(Buff);
 
-	lower_bar();
+	lower_bar(Buff);
 	set_cur_pos(Buff, Ui);
 }
 
-void set_cur_pos(f_mtdt* Buff, ui_mtdt Ui)
+void set_cur_pos(f_mtdt* Buff, win_mtdt Ui)
 {
 	// Case when all lines fits in the window.
 	term_t mv_up = 0;