+++ src/render.h
@@ -19,16 +19,14 @@
 "CTRL^: X - exit; W - save; G/H - cursor; S/Q - render/hide new chars\0"
 
 // Remember to not override the upper bar width.
-#define TERM_X_MIN (term_t) strlen(LBAR_STR)
+#define TERM_X_MIN (term_t) (strlen(LBAR_STR) + SPACE_SZ)
 
 // Sizes of the text area.
 #define TXT_Y (term_t) (get_term_sz(Buff, 'Y') - BARS_SZ) // TODO: IN UI STRUCT.
 
 // ANSI escape codes:
 #define ANSI_RESET()           printf("\033[%s", "0m")
-
-// If you want to change to inverted bars' colors, set "1m" to "7m".
-#define ANSI_BOLD()            printf("\033[%s", "1m")
+#define ANSI_INVERT()          printf("\033[%s", "7m")
 #define ANSI_CLEAN_LN()        printf("\033[%s", "2K")
 #define ANSI_CUR_UP(offset)    printf("\033[%dA", offset)
 #define ANSI_CUR_DOWN(offset)  printf("\033[%dB", offset)
@@ -44,7 +42,7 @@ typedef struct
 	uint8_t line_num_len;     // Dynamic width of the lines numbers.
 	term_t  text_x;           // Horizontal space for the text (width: chars).
 }
-ui_mtdt;
+win_mtdt;
 #pragma pack(pop)
 
 // Returns current terminal width and height and exits if is wrong.
@@ -54,13 +52,13 @@ term_t get_term_sz(f_mtdt* Buff, char axis);
 void flush_window(f_mtdt* Buff);
 
 // Renders the upper bar with a filename and indicators.
-void upper_bar(f_mtdt* Buff, ui_mtdt Ui);
+void upper_bar(f_mtdt* Buff, win_mtdt Ui);
 
 // Renders the lower bar that contains keyboard info.
-void lower_bar(void);
+void lower_bar(f_mtdt* Buff);
 
 // Scrolls chars. Used when the cursor is in static position.
-void scroll_line_x(f_mtdt* Buff, ui_mtdt Ui);
+void scroll_line_x(f_mtdt* Buff, win_mtdt Ui);
 
 // Returns value of hidden lines.
 buff_t scroll_lines(f_mtdt* Buff);
@@ -69,7 +67,7 @@ buff_t scroll_lines(f_mtdt* Buff);
 void print_line_num(buff_t line, uint8_t line_num_len);
 
 // Shows a text in the window.
-void display_text(f_mtdt* Buff, ui_mtdt Ui);
+void display_text(f_mtdt* Buff, win_mtdt Ui);
 
 // Vertical fill between the text and lower bar. If there isn't many lines.
 void fill(f_mtdt* Buff);
@@ -78,7 +76,7 @@ void fill(f_mtdt* Buff);
 void window(f_mtdt* Buff);
 
 // Sets the cursor position from the left bottom.
-void set_cur_pos(f_mtdt* Buff, ui_mtdt Ui);
+void set_cur_pos(f_mtdt* Buff, win_mtdt Ui);
 
 #endif